% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/edit_rprofile.R
\name{edit_rprofile.site}
\alias{edit_rprofile.site}
\title{Edit the Rprofile.site file}
\usage{
edit_rprofile.site(create_file = FALSE)
}
\arguments{
\item{create_file}{boolean indicating if the Rprofile.site file should be created if it doesn't exist.}
}
\description{
Function to edit the Rprofile.site file
}
