% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/04_tools.R
\name{fd}
\alias{fd}
\title{Forward Difference}
\usage{
fd(x)
}
\arguments{
\item{x}{numeric vector or matrix.}
}
\value{
an atomic vector or matrix of the same dimensions as \code{x}.
}
\description{
Calculates the forward difference of a vector or matrix.
}
\details{
The function is used internally in the \code{Multigam} distribution.
}
\examples{
\dontrun{
set.seed(1)
x <- r(Multigam(1:3, 1), 4)
fd(x)
}
}
\keyword{internal}
