% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/04_tools.R
\name{inv2x2}
\alias{inv2x2}
\title{2x2 Inverse Matrix}
\usage{
inv2x2(x)
}
\arguments{
\item{x}{A 2x2 matrix.}
}
\value{
A 2x2 matrix, the inverse of \code{x}
}
\description{
Calculates the inverse of a 2x2 matrix.
}
\examples{
\dontrun{
x <- matrix(1:4, nrow = 2, ncol = 2)
inv2x2(x)
}
}
\keyword{internal}
