% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/random_joke.R
\name{joke_categories}
\alias{joke_categories}
\title{Joke categories}
\usage{
joke_categories()
}
\description{
Function to get the different categories to filter random jokes
}
\examples{
joke_categories()
}
