% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/joker-package.R
\docType{package}
\name{joker-package}
\alias{joker}
\alias{joker-package}
\title{joker: Probability Distributions and Parameter Estimation}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

Implements an S4 distribution system and estimation methods for parameters of common distribution families. The common d, p, q, r function family for each distribution is enriched with the ll, e, and v counterparts, computing the log-likelihood, performing estimation, and calculating the asymptotic variance - covariance matrix, respectively. Parameter estimation is performed analytically whenever possible.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://thechibo.github.io/joker}
  \item Report bugs at \url{https://github.com/thechibo/joker/issues}
}

}
\author{
\strong{Maintainer}: Ioannis Oikonomidis \email{goikon@math.uoa.gr} (\href{https://orcid.org/0000-0001-8130-2104}{ORCID})

Authors:
\itemize{
  \item Samis Trevezas \email{strevezas@math.uoa.gr} (\href{https://orcid.org/0000-0003-2262-8299}{ORCID}) [thesis advisor]
}

}
\keyword{internal}
