% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colors.R
\name{joker}
\alias{joker}
\title{Joker palette generator}
\usage{
joker(name, n)
}
\arguments{
\item{name}{Name of desired palette. Choices are:
\code{joker_2019}}

\item{n}{Number of colors desired.}
}
\value{
A vector of colors.
}
\description{
Joker palette generator
}
\examples{
joker("joker_2019")
joker("joker_2019", 3)
}
