% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/knock_knock.R
\name{knock_knock}
\alias{knock_knock}
\title{Returns a Knock-Knock joke}
\usage{
knock_knock()
}
\description{
Returns a random knock-knock joke of the 300+ knock-knock jokes pre-installed.
}
\examples{

knock_knock()
  
}
