% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/04_tools.R
\name{Matrix}
\alias{Matrix}
\alias{nearPD}
\title{Matrix Wrappers}
\usage{
Matrix(...)

nearPD(x)
}
\arguments{
\item{...}{extra arguments passed to \code{Matrix::Matrix()}.}

\item{x}{matrix}
}
\value{
matrix
}
\description{
Simple wrappers of functions from the \code{Matrix} package.
}
\keyword{internal}
