% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/04_tools.R
\name{progress}
\alias{progress}
\alias{progress_bar}
\alias{format_hms}
\title{Progress Bar}
\usage{
progress_bar(iter, total, start, message = NULL, width = 20)

format_hms(seconds)
}
\arguments{
\item{iter}{integer. The current iteration step of the for loop.}

\item{total}{integer. The total number of iterations.}

\item{start}{POSIXct. The start time, as returned by \code{Sys.time()}.}

\item{message}{character. A message appearing before the progress bar.}

\item{width}{integer. The length of the progress bar.}

\item{seconds}{integer. Seconds to be converted into hh:mm:ss format.}
}
\value{
\code{format_hms()} returns a character in the hh:mm:ss format. \code{progress_bar()}
prints the progress bar on the console, calling \code{cat()}, therefore as returns
an invisible \code{NULL}.
}
\description{
Create a progress bar to be used with for loops that can possibly take a lot
of time.
}
\examples{
\dontrun{
# Turn seconds into hh:mm:ss
format_hms(7543)

n <- 100
start <- Sys.time()

for (i in 1:n) {
  Sys.sleep(0.1)  # Simulate time-consuming task
  progress_bar(i, n, start, message = "Computing", width = 30)
}
}
}
\keyword{internal}
