% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/random_joke.R
\name{random_joke}
\alias{random_joke}
\title{Random joke}
\usage{
random_joke(data = c("stupidstuff.json", "wocka.json"))
}
\arguments{
\item{data}{database to use}
}
\description{
Function to print a random joke from a database
}
\examples{
random_joke()
}
