% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/04_tools.R
\name{set1of1}
\alias{set1of1}
\alias{set1of2}
\alias{set1of3}
\alias{set2of3}
\title{Indexing Functions}
\usage{
set1of1(x, i)

set1of2(x, i)

set1of3(x, i)

set2of3(x, i)
}
\arguments{
\item{x}{atomic vector, matrix, or array. An object to be indexed.}

\item{i}{integer. The index.}
}
\value{
A vector or matrix, subset of the original object.
}
\description{
A set of functions that allow to index a matrix or array. These functions are
used internally when the dimension length of an object can vary.
}
\examples{
\dontrun{
x <- 1:12
y <- matrix(1:12, nrow = 3, ncol = 4)
z <- array(1:24, dim = c(3, 4, 2))

set1of1(x)
set1of2(y)
set1of3(z)
set2of3(z)
}
}
\keyword{internal}
