% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/04_tools.R
\name{stats}
\alias{stats}
\alias{bvar}
\alias{bsd}
\alias{rowVar}
\alias{colVar}
\alias{skew,numeric-method}
\alias{kurt,numeric-method}
\title{Statistical Functions}
\usage{
bvar(x)

bsd(x)

rowVar(x)

colVar(x)

\S4method{skew}{numeric}(x)

\S4method{kurt}{numeric}(x)
}
\arguments{
\item{x}{numeric for \code{bvar()} and \code{bsd()}, matrix for \code{rowVar()} and
\code{colVar()}.}
}
\value{
\code{bvar()} and \code{bsd()} return a single numeric, \code{rowVar()} and
\code{colVar()} return a numeric vector.
}
\description{
A set of statistics that extend the ones made available in the \code{stats}
package.
}
\section{Functions}{
\itemize{
\item \code{bvar()}: Biased sample variance

\item \code{bsd()}: Biased sample standard deviation

\item \code{rowVar()}: Biased sample variance by matrix row

\item \code{colVar()}: Biased sample variance by matrix column

\item \code{skew(numeric)}: Sample skewness

\item \code{kurt(numeric)}: Sample kurtosis

}}
\examples{
\dontrun{
x <- r(Beta(2, 3), 20)
y <- matrix(x, nrow = 2, ncol = 10)

bvar(x)
bsd(x)
rowVar(y)
rowCol(y)
}
}
\keyword{internal}
