% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TellJokeAbout.R
\name{TellJokeAbout}
\alias{TellJokeAbout}
\title{TellJokeAbout}
\usage{
TellJokeAbout(topic)
}
\arguments{
\item{topic}{A character value. Topic which the joke is a about}
}
\value{
A character value. A sentence with a joke
}
\description{
Tells a joke about a Eoghan.
}
\seealso{
\url{https://icanhazdadjoke.com}
}
\author{
Gareth Burns
}
