% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TellRandomJoke.R
\name{TellRandomJoke}
\alias{TellRandomJoke}
\title{TellRandomJoke}
\usage{
TellRandomJoke(...)
}
\value{
A character value. A sentence with a joke
}
\description{
Tells a random joke
}
\seealso{
\url{https://icanhazdadjoke.com}
}
\author{
Gareth Burns
}
