% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/04_tools.R
\name{array_to_df}
\alias{array_to_df}
\title{Turn an array to a data.frame}
\usage{
array_to_df(x)
}
\arguments{
\item{x}{array.}
}
\value{
data.frame.
}
\description{
This function turns an array to a data.frame. It is used by the
\code{small_metrics()} and \code{large_metrics()} functions.
}
\keyword{internal}
