% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02_Cauchy.R
\name{Cauchy}
\alias{Cauchy}
\alias{d,Cauchy,numeric-method}
\alias{p,Cauchy,numeric-method}
\alias{qn,Cauchy,numeric-method}
\alias{r,Cauchy,numeric-method}
\alias{mean,Cauchy-method}
\alias{median,Cauchy-method}
\alias{mode,Cauchy-method}
\alias{var,Cauchy-method}
\alias{sd,Cauchy-method}
\alias{skew,Cauchy-method}
\alias{kurt,Cauchy-method}
\alias{entro,Cauchy-method}
\alias{finf,Cauchy-method}
\alias{llcauchy}
\alias{ll,Cauchy,numeric-method}
\alias{ecauchy}
\alias{mle,Cauchy,numeric-method}
\alias{me,Cauchy,numeric-method}
\alias{vcauchy}
\alias{avar_mle,Cauchy-method}
\title{Cauchy Distribution}
\usage{
Cauchy(location = 0, scale = 1)

\S4method{d}{Cauchy,numeric}(distr, x)

\S4method{p}{Cauchy,numeric}(distr, x)

\S4method{qn}{Cauchy,numeric}(distr, x)

\S4method{r}{Cauchy,numeric}(distr, n)

\S4method{mean}{Cauchy}(x)

\S4method{median}{Cauchy}(x)

\S4method{mode}{Cauchy}(x)

\S4method{var}{Cauchy}(x)

\S4method{sd}{Cauchy}(x)

\S4method{skew}{Cauchy}(x)

\S4method{kurt}{Cauchy}(x)

\S4method{entro}{Cauchy}(x)

\S4method{finf}{Cauchy}(x)

llcauchy(x, location, scale)

\S4method{ll}{Cauchy,numeric}(distr, x)

ecauchy(x, type = "mle", ...)

\S4method{mle}{Cauchy,numeric}(
  distr,
  x,
  par0 = "me",
  method = "L-BFGS-B",
  lower = c(-Inf, 1e-05),
  upper = c(Inf, Inf)
)

\S4method{me}{Cauchy,numeric}(distr, x)

vcauchy(location, scale, type = "mle")

\S4method{avar_mle}{Cauchy}(distr)
}
\arguments{
\item{location, scale}{numeric. The distribution parameters.}

\item{distr, x}{If both arguments coexist, \code{distr} is an object of class
\code{Cauchy} and \code{x} is a numeric vector, the sample of observations. For the
moment functions that only take an \code{x} argument, \code{x} is an object of class
\code{Cauchy} instead.}

\item{n}{numeric. The sample size.}

\item{type}{character, case ignored. The estimator type (mle, me, or same).}

\item{...}{extra arguments.}

\item{par0, method, lower, upper}{arguments passed to optim for the mle
optimization.}
}
\value{
Each type of function returns a different type of object:
\itemize{
\item Distribution Functions: When supplied with one argument (\code{distr}), the
\code{d()}, \code{p()}, \code{q()}, \code{r()}, \code{ll()} functions return the density, cumulative
probability, quantile, random sample generator, and log-likelihood functions,
respectively. When supplied with both arguments (\code{distr} and \code{x}), they
evaluate the aforementioned functions directly.
\item Moments: Returns a numeric, either vector or matrix depending on the moment
and the distribution. The \code{moments()} function returns a list with all the
available methods.
\item Estimation: Returns a list, the estimators of the unknown parameters. Note
that in distribution families like the binomial, multinomial, and negative
binomial, the size is not returned, since it is considered known.
\item Variance: Returns a named matrix. The asymptotic covariance matrix of the
estimator.
}
}
\description{
The Cauchy distribution is an absolute continuous probability distribution
characterized by its location parameter \eqn{x_0} and scale parameter
\eqn{\gamma > 0}.
}
\details{
The probability density function (PDF) of the Cauchy distribution is given
by: \deqn{ f(x; x_0, \gamma) = \frac{1}{\pi \gamma \left[1 + \left(\frac{x - x_0}{\gamma}\right)^2\right]}.}
}
\examples{
# -----------------------------------------------------
# Cauchy Distribution Example
# -----------------------------------------------------

# Create the distribution
x0 <- 3 ; scale <- 5
D <- Cauchy(x0, scale)
x <- c(-5, 3, 10)
n <- 100

# ------------------
# dpqr Functions
# ------------------

d(D, x) # density function
p(D, x) # distribution function
qn(D, 0.8) # inverse distribution function
x <- r(D, n) # random generator function

# alternative way to use the function
df <- d(D) ; df(x) # df is a function itself

# ------------------
# Moments
# ------------------

median(D) # Median
mode(D) # Mode
entro(D) # Entropy
finf(D) # Fisher Information Matrix

# ------------------
# Point Estimation
# ------------------

ll(D, x)
llcauchy(x, x0, scale)

ecauchy(x, type = "mle")
ecauchy(x, type = "me")

mle(D, x)
me(D, x)
e(D, x, type = "mle")

mle("cauchy", x) # the distr argument can be a character

# ------------------
# As. Variance
# ------------------

vcauchy(x0, scale, type = "mle")
avar_mle(D)
avar(D, type = "mle")
}
\seealso{
Functions from the \code{stats} package: \code{\link[=dcauchy]{dcauchy()}}, \code{\link[=pcauchy]{pcauchy()}}, \code{\link[=qcauchy]{qcauchy()}}, \code{\link[=rcauchy]{rcauchy()}}
}
