% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/04_tools.R
\name{check_data}
\alias{check_data}
\title{Check the Data}
\usage{
check_data(x, na.rm = FALSE)
}
\arguments{
\item{x}{an object to be checked. Should be vector, matrix, or array.}

\item{na.rm}{logical. Should the NA values be removed?}
}
\value{
The object \code{x}, possibly without \code{NA} values if \code{x} is a vector
containing \code{NA} values and \code{na.rm = TRUE}.
}
\description{
This function checks that the data argument supplied by the user for
parameter estimation are of the appropriate type.
}
\examples{
\dontrun{
check_data(r(Beta(2, 3), 20))
check_data(c(1, 2, NA))
check_data(c(1, 2, NA), na.rm = TRUE)
check_data(c(1, 2, NaN))
check_data(c(1, 2, Inf))
}
}
\keyword{internal}
