% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/04_tools.R
\name{set1of1}
\alias{set1of1}
\alias{set1of2}
\alias{set1of3}
\alias{set2of3}
\title{Indexing Functions}
\usage{
set1of1(x, i)

set1of2(x, i)

set1of3(x, i)

set2of3(x, i)
}
\arguments{
\item{x}{atomic vector, matrix, or array. An object to be indexed.}

\item{i}{integer. The index.}
}
\value{
A vector or matrix, subset of the original object.
}
\description{
A set of functions that allow to index a matrix or array. These functions are
used internally when the dimension length of an object can vary.
}
\keyword{internal}
